moved controllers to their own assembly and fixed depedency issues
