+++ aspnet/SDP.TaskManagement.Web/SDP.TaskManagement.Web.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="9.0.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\SDP.TaskManagement.Application\SDP.TaskManagement.Application.csproj" />
+    <ProjectReference Include="..\src\SDP.TaskManagement.Domain\SDP.TaskManagement.Domain.csproj" />
+    <ProjectReference Include="..\src\SDP.TaskManagement.Infrastructure\SDP.TaskManagement.Infrastructure.csproj" />
+  </ItemGroup>
+
+</Project>
