+++ aspnet/src/SDP.TaskManagement.WebHost/SDP.TaskManagement.WebHost.csproj
@@ -17,6 +17,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <ProjectReference Include="..\..\SDP.TaskManagement.Web\SDP.TaskManagement.Web.csproj" />
 	  <ProjectReference Include="..\SDP.TaskManagement.Application\SDP.TaskManagement.Application.csproj" />
 	  <ProjectReference Include="..\SDP.TaskManagement.Infrastructure\SDP.TaskManagement.Infrastructure.csproj" />
 	</ItemGroup>
